backend/requirements.txt
Django==3.2.16
djangorestframework==3.12.4
Pillow==9.3.0
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
django-filter==23.5
